#51 Move services and utils to common folder
