+++ ReportJ.Extension.Chrome/app/js/common/globals.js
