+++ ReportJ.Extension.Chrome/app/js/common/utils/browser.js
