+++ ReportJ.Extension.Chrome/app/js/common/utils/dialog.js
