+++ ReportJ.Extension.Chrome/app/js/common/utils/function.js
