+++ ReportJ.Extension.Chrome/app/js/common/utils/html-observers/elementObserver.js
