+++ ReportJ.Extension.Chrome/app/js/common/utils/jira.js
