+++ ReportJ.Extension.Chrome/app/js/common/utils/object.js
