+++ ReportJ.Extension.Chrome/app/js/common/utils/react.js
