+++ ReportJ.Extension.Chrome/app/js/common/utils/url.js
