+++ ReportJ.Extension.Chrome/app/js/options/app.js
@@ -1,6 +1,6 @@
 import OptionsController from "./controllers/optionsController";
-import RepositoryService from "~/js/services/repositoryService";
-import StarageService from "~/js/services/storageService";
+import RepositoryService from "app/js/common/services/repositoryService";
+import StarageService from "app/js/common/services/storageService";
 import angular from "angular";
 import diableValidation from "./directives/diableValidation";
 import equalsDirective from "./directives/equalsDirective";
