+++ ReportJ.Extension.Chrome/config/base.config.js
@@ -33,10 +33,10 @@ function getSassLoaders() {
 module.exports = function () {
     return {
         entry: {
-            background: "./app/js/background.js",
-            popup: "./app/js/popup/index.jsx",
-            options: "./app/js/options/app.js",
-            content: "./app/js/content/content.js"
+            background: "./app/js/background/background.js",
+            popup:      "./app/js/popup/index.jsx",
+            options:    "./app/js/options/app.js",
+            content:    "./app/js/content/content.js"
         },
         output: {
             filename: "[name].js",
